bug fix - after reseting sort on table an error was thrown
there was a problem that afer reseting sort, the $request->sortBy array did not have the key "direction". Switched up the if else, and added additional safeguard check if the key exists, now we can reset the sort and change pagination pages normaly. Before on pagination change an error was thrown.